fix: updated program style
Moved * closer to variables. This better fit the syntax of the language since the pointer operator is part of the variable and not part of the type. In fact, if you would need to declare two pointer to integers you would put the * twice even though you can define the type only once: int *i, j*;